+++ lib/slather/command/version_command.rb
@@ -1,0 +1,6 @@
+class VersionCommand < Clamp::Command
+
+  def execute
+    puts "slather #{Slather::VERSION}"
+  end
+end